requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
better-exceptions==0.3.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
uuid==1.30
psycopg2-binary==2.9.9
aiofiles==23.2.1
cachetools==5.3.2